#13 extracting footer styles to separate page and removing unused sections of code
